web7

+++ delete.php
@@ -5,9 +5,12 @@ try {
     $password = '1377191';
     $db = new PDO('mysql:host=localhost;dbname=u20296', $user, $password);
     $login = $_POST['sendform'];
-    $sth = $db->prepare("DELETE FROM DBlab5 WHERE login=:login");
-    $sth->bindParam(':login', $login);
-    $sth->execute();
+    $token = $_POST['token'];
+    if ($token=="cant_be_hacked"){
+        $sth = $db->prepare("DELETE FROM DBlab5 WHERE login=:login");
+        $sth->bindParam(':login', $login);
+        $sth->execute();
+    }
     header('Location: admin.php');
 }
 catch(PDOException $e){
